requirements.txt
numpy==1.16.2
scikit-learn==0.20.1
scipy==1.2.1
sklearn==0.0
beautifulsoup4==4.6.1
google-api-python-client==1.7.4
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
